Switch Razor LSP server code to use RazorFormattingService from common layer
