--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Formatting/FormattingSpanKind.cs
@@ -1,14 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
-
-internal enum FormattingSpanKind
-{
-    Transition,
-    MetaCode,
-    Comment,
-    Code,
-    Markup,
-    None
-}
