--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Formatting/IFormattingPass.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
-
-internal interface IFormattingPass
-{
-    int Order { get; }
-
-    bool IsValidationPass { get; }
-
-    Task<FormattingResult> ExecuteAsync(FormattingContext context, FormattingResult result, CancellationToken cancellationToken);
-}
