+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Formatting/RazorFormattingPassBase.cs
@@ -6,6 +6,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Razor;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Components;
 using Microsoft.AspNetCore.Razor.Language.Extensions;
@@ -14,15 +15,14 @@ using Microsoft.CodeAnalysis.Razor.DocumentMapping;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
+namespace Microsoft.CodeAnalysis.Razor.Formatting;
 
-internal sealed class RazorFormattingPass(
-    IDocumentMappingService documentMappingService,
-    RazorLSPOptionsMonitor optionsMonitor)
+using SyntaxNode = AspNetCore.Razor.Language.Syntax.SyntaxNode;
+
+internal abstract class RazorFormattingPassBase(
+    IDocumentMappingService documentMappingService)
     : FormattingPassBase(documentMappingService)
 {
-    private readonly RazorLSPOptionsMonitor _optionsMonitor = optionsMonitor;
-
     // Run after the C# formatter pass.
     public override int Order => DefaultOrder - 4;
 
@@ -63,6 +63,8 @@ internal sealed class RazorFormattingPass(
         return new FormattingResult(finalEdits);
     }
 
+    protected abstract bool CodeBlockBraceOnNextLine { get; }
+
     private IEnumerable<TextEdit> FormatRazor(FormattingContext context, RazorSyntaxTree syntaxTree)
     {
         var edits = new List<TextEdit>();
@@ -72,7 +74,7 @@ internal sealed class RazorFormattingPass(
         {
             // Disclaimer: CSharpCodeBlockSyntax is used a _lot_ in razor so these methods are probably
             // being overly careful to only try to format syntax forms they care about.
-            TryFormatCSharpBlockStructure(context, edits, source, node);
+            TryFormatCSharpBlockStructure(context, edits, source, node, codeBlockBraceOnNextLine: true); // TODO
             TryFormatSingleLineDirective(edits, source, node);
             TryFormatBlocks(context, edits, source, node);
         }
@@ -126,7 +128,7 @@ internal sealed class RazorFormattingPass(
 
         return false;
 
-        static bool TryGetWhitespace(SyntaxList<RazorSyntaxNode> children, [NotNullWhen(true)] out CSharpStatementLiteralSyntax? whitespaceBeforeSectionName, [NotNullWhen(true)] out UnclassifiedTextLiteralSyntax? whitespaceAfterSectionName)
+        static bool TryGetWhitespace(AspNetCore.Razor.Language.Syntax.SyntaxList<RazorSyntaxNode> children, [NotNullWhen(true)] out CSharpStatementLiteralSyntax? whitespaceBeforeSectionName, [NotNullWhen(true)] out UnclassifiedTextLiteralSyntax? whitespaceAfterSectionName)
         {
             // If there is whitespace between the directive and the section name, and the section name and the brace, they will be in the first child
             // and third child of the 6 total children
@@ -246,7 +248,7 @@ internal sealed class RazorFormattingPass(
         return false;
     }
 
-    private void TryFormatCSharpBlockStructure(FormattingContext context, List<TextEdit> edits, RazorSourceDocument source, SyntaxNode node)
+    private void TryFormatCSharpBlockStructure(FormattingContext context, List<TextEdit> edits, RazorSourceDocument source, SyntaxNode node, bool codeBlockBraceOnNextLine)
     {
         // We're looking for a code block like this:
         //
@@ -266,7 +268,7 @@ internal sealed class RazorFormattingPass(
             directive.DirectiveDescriptor?.Kind == DirectiveKind.CodeBlock)
         {
             // If we're formatting a @code or @functions directive, the user might have indicated they always want a newline
-            var forceNewLine = _optionsMonitor.CurrentValue.CodeBlockBraceOnNextLine &&
+            var forceNewLine = codeBlockBraceOnNextLine &&
                 directive.Body is RazorDirectiveBodySyntax { Keyword: { } keyword } &&
                 IsCodeOrFunctionsBlock(keyword);
 
@@ -290,7 +292,7 @@ internal sealed class RazorFormattingPass(
             }
         }
 
-        static bool TryGetLeadingWhitespace(SyntaxList<RazorSyntaxNode> children, [NotNullWhen(true)] out UnclassifiedTextLiteralSyntax? whitespace)
+        static bool TryGetLeadingWhitespace(AspNetCore.Razor.Language.Syntax.SyntaxList<RazorSyntaxNode> children, [NotNullWhen(true)] out UnclassifiedTextLiteralSyntax? whitespace)
         {
             // If there is whitespace between the directive and the brace, it will be in the first child
             // of the 4 total children
@@ -327,7 +329,7 @@ internal sealed class RazorFormattingPass(
             }
         }
 
-        static bool IsSingleLineDirective(SyntaxNode node, out SyntaxList<SyntaxNode> children)
+        static bool IsSingleLineDirective(SyntaxNode node, out AspNetCore.Razor.Language.Syntax.SyntaxList<SyntaxNode> children)
         {
             if (node is CSharpCodeBlockSyntax content &&
                 node.Parent?.Parent is RazorDirectiveSyntax directive &&
