+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/IAdhocWorkspaceFactory.cs
@@ -1,10 +1,9 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Host;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer;
+namespace Microsoft.CodeAnalysis.Razor.Workspaces;
 
 internal interface IAdhocWorkspaceFactory
 {
